requirements.txt.txt
streamlit
google-generativeai
biopython
fpdf
plotly
pandas